implement new metric for individual requests duration
Add a metric to measure actual latency of every requests. This new
metric will help troubleshooting and investigations as it shows the
evolution of requests duration over time.
Note that this new metric's name `proxy_request_time_seconds` is very
similar to `proxy_requests_time_seconds` but they are not exactly the
same. You can always refer to the Prometheus help in the code for more details.